requirements.txt
# 核心套件
streamlit>=1.28.0
insightface>=0.7.3
onnxruntime>=1.16.0

# 影像處理
opencv-python>=4.8.0
Pillow>=10.0.0
matplotlib>=3.7.0

# 數據處理
numpy>=1.24.0
pandas>=2.0.0

# 其他依賴
scikit-learn>=1.3.0
